[refactor] :truck: add suffix to all types (#276)
