+++ src/services/manteca/transactions/mantecaFiatTrxsService.ts
@@ -1,8 +1,8 @@
 import axios from 'axios';
 
 import { Logger } from '../../../helpers/loggerHelper';
-import { MantecaTransaction } from '../../../types/manteca';
 import { MANTECA_BASE_URL } from '../../../config/constants';
+import { MantecaTransaction } from '../../../types/mantecaType';
 import { getMantecaAxiosConfig } from '../mantecaCommonService';
 
 export const mantecaFiatTrxsService = {
