+++ src/types/chatizaloType.ts
@@ -1,0 +1,5 @@
+export interface chatizaloOperatorReplyType {
+  data_token: string;
+  channel_user_id: string;
+  message: string;
+}
